+++ grails-app/controllers/io/xh/hoist/RestController.groovy
@@ -9,6 +9,7 @@ package io.xh.hoist
 
 import grails.validation.ValidationException
 import io.xh.hoist.json.JSONParser
+import static io.xh.hoist.util.Utils.handleException
 
 abstract class RestController extends BaseController {
 
@@ -72,7 +73,7 @@ abstract class RestController extends BaseController {
                     successCount++
                 } catch (Exception e) {
                     failCount++
-                    xhExceptionHandler.handleException(
+                    handleException(
                         exception: e,
                         logMessage: "Exception updating ${id}",
                         logTo: this