+++ grails-app/controllers/io/xh/hoist/admin/cluster/LogViewerAdminController.groovy
@@ -12,6 +12,7 @@ import io.xh.hoist.BaseController
 import io.xh.hoist.security.Access
 import static io.xh.hoist.util.ClusterUtils.runOnInstanceAsJson
 import static io.xh.hoist.util.ClusterUtils.runOnInstance
+import static io.xh.hoist.util.Utils.handleException
 
 
 @Access(['HOIST_ADMIN_READER'])
@@ -46,7 +47,7 @@ class LogViewerAdminController extends BaseController {
 
         if (ret.exception) {
             // Just render exception, was already logged on target instance
-            xhExceptionHandler.handleException(exception: ret.exception, renderTo: response)
+            handleException(exception: ret.exception, renderTo: response)
             return
         }
         def file = ret.value as File